refactor: Exclude `Tags` interface from `TraceSpan` context
