+++ node/packages/sdk/test/unit/lib/trace-span.test.js
@@ -178,53 +178,4 @@ describe('lib/trace-span.test.js', () => {
     expect(Long.isLong(protoJson.startTimeUnixNano)).to.be.true;
     expect(Long.isLong(protoJson.endTimeUnixNano)).to.be.true;
   });
-
-  describe('tags', () => {
-    it('should allow setting tags', () => {
-      const childSpan = new TraceSpan('child').close();
-      childSpan.tags.set('bool', true);
-      childSpan.tags.set('string', 'string');
-      childSpan.tags.set('num', 23);
-      childSpan.tags.set('strings', ['foo', 'bar']);
-      childSpan.tags.set('numbers', [1, 2]);
-      expect(Array.from(childSpan.tags)).to.deep.equal([
-        ['bool', true],
-        ['string', 'string'],
-        ['num', 23],
-        ['strings', ['foo', 'bar']],
-        ['numbers', [1, 2]],
-      ]);
-    });
-
-    it('should support setting many tags at once', () => {
-      const childSpan = new TraceSpan('child').close();
-      childSpan.tags.setMany({ bool: true, string: 'string', num: 23, novalue: null });
-      childSpan.tags.setMany(
-        { bool: true, string: 'string', num: 23, novalue: null },
-        { prefix: 'foo.bar' }
-      );
-      expect(Array.from(childSpan.tags)).to.deep.equal([
-        ['bool', true],
-        ['string', 'string'],
-        ['num', 23],
-        ['foo.bar.bool', true],
-        ['foo.bar.string', 'string'],
-        ['foo.bar.num', 23],
-      ]);
-    });
-
-    it('should reject setting with different value tag that alraedy exists', () => {
-      const childSpan = new TraceSpan('child').close();
-      childSpan.tags.set('tag', true);
-      expect(() => childSpan.tags.set('tag', 'again'))
-        .to.throw(Error)
-        .with.property('code', 'DUPLICATE_TRACE_SPAN_TAG_NAME');
-    });
-
-    it('should ignore resetting tag with same value', () => {
-      const childSpan = new TraceSpan('child').close();
-      childSpan.tags.set('tag', true);
-      expect(childSpan.tags.set('tag', true)).to.equal(childSpan.tags);
-    });
-  });
 });
